menambahkan fungction untuk register dan authentikasi user

+++ base/views.py
@@ -1,14 +1,20 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib import messages
 from django.contrib.auth import login,logout,authenticate
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
 
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect("home")
 
     if request.method == "POST":
         username = request.POST.get("username")
@@ -27,12 +33,33 @@ def loginPage(request):
             messages.error(request, "Username Or Passoword doesn't exist")
 
 
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login.html' ,context)
 
 def logoutUser(request):
     logout(request)
     return redirect("home")
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Kesalahan saat pendaftaran")
+
+
+    context = {
+        'form':form
+    }
+    return render(request, 'base/login.html', context)
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -56,6 +83,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm
     if request.method == "POST":
@@ -66,9 +94,12 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+            return HttpResponse('You are not authorized to edit this room.')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -79,8 +110,11 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+            return HttpResponse('You are not authorized to edit this room.')
     if request.method == "POST":
         room.delete()
         return redirect("home")
